requirements.txt
# ── Core ──────────────────────────────────────────────────────────────────────
pydantic==2.5.0
pyyaml==6.0.1
python-dotenv==1.0.0
tenacity==8.2.3
click==8.1.7

# ── API ───────────────────────────────────────────────────────────────────────
fastapi==0.108.0
uvicorn[standard]==0.25.0
grpcio==1.60.0
grpcio-tools==1.60.0
httpx==0.26.0                   # async HTTP client + TestClient

# ── Streaming ─────────────────────────────────────────────────────────────────
kafka-python==2.0.2
confluent-kafka==2.3.0

# ── Storage ───────────────────────────────────────────────────────────────────
redis==5.0.1
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
pyarrow==14.0.0

# ── Data Processing ───────────────────────────────────────────────────────────
pandas==2.1.4
numpy==1.26.2
polars==0.20.3
# pyspark==3.5.0  # uncomment for Spark support (requires Java)

# ── Machine Learning ──────────────────────────────────────────────────────────
scikit-learn==1.3.2
joblib==1.3.2

# ── Monitoring ────────────────────────────────────────────────────────────────
prometheus-client==0.19.0
python-json-logger==2.0.7

# ── Testing ───────────────────────────────────────────────────────────────────
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-benchmark==4.0.0
faker==21.0.0
pytest-cov==4.1.0

# ── Dev tooling ───────────────────────────────────────────────────────────────
black==23.12.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0
pre-commit==3.6.0
